Refactor: Standardize domain entity naming and audit properties
- Renamed entities to use singular names (Applications -> Application, Counties -> Country, LicenseClasses -> LicenseClass, TestTypes -> TestType, ApplicationTypes -> ApplicationType).

- Removed audit and deactivation properties (CreatedAt, UpdatedAt, IsDeactivated) from the base Entity class.

- Added explicit CreateAt and UpdateAt audit properties to individual domain entities.

- Refactored domain logic and relationships to reference the new singular entity types.

- Updated User entity to use Role enum instead of a string property.

+++ backend/DVLD.Domain/Enums/Gender.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace DVLD.Domain.Enums
 {
-    public enum Gender : short
+    public enum Gender : byte
     {
         Male = 1,
         Female = 2,
